Relocate 7 verifier message helpers to verifier_orchestration
Phase 2 follow-up PR for Issue #682. Moves 7 small/medium pure-fn
verifier helpers (~46 LOC of bodies) out of turn.rs into
verifier_orchestration.rs, alongside the 7 type definitions migrated
in #714.

Moved (turn.rs -> verifier_orchestration.rs):

  Verifier repair policy messages (pure const strings):
    - verifier_repair_transition_message
    - verifier_repair_safe_stop_message
    - verifier_repair_unsafe_target_message

  Error / status message factories:
    - verifier_repair_pass_request_error_message
    - task_contract_verifier_target_discovery_note (#[cfg(test)] only)

  Display / attempt-limit helpers:
    - verifier_repair_target_display
    - task_contract_verifier_failure_attempt_limit

All 7 are now `pub(super)` (most were file-private before).
task_contract_verifier_target_discovery_note keeps `#[cfg(test)]`
since it was test-only in turn.rs; the call site (mod progress_tests)
imports it explicitly from super::super::verifier_orchestration::.

Adjustments:

  * verifier_orchestration.rs: added
    `use std::path::Path;` and
    `use super::turn::{TASK_CONTRACT_VERIFIER_ATTEMPT_LIMIT,
    TASK_CONTRACT_VERIFIER_REPAIR_ATTEMPT_LIMIT};` back-imports for the
    two consts that the moved fn references; the consts stay in
    turn.rs (still consumed by other turn.rs callers).

  * turn.rs:
    - Promoted TASK_CONTRACT_VERIFIER_REPAIR_ATTEMPT_LIMIT from
      file-private to pub(super) so verifier_orchestration.rs can
      import it.
    - Extended `use super::verifier_orchestration::{...}` to pull
      the 6 production-callable fns back in.
    - Dropped task_contract_verifier_target_discovery_note from
      mod tests's `use super::{...}` (now imported from the new
      module in mod progress_tests directly).
    - Removed an orphaned `#[cfg(test)]` attribute that was left
      attached to verifier_repair_diagnostic_pending_note after
      target_discovery_note was moved out (it would otherwise
      cfg-gate the next function out of release builds).
    - Restored 7 language-pattern consts (PYTHON_REQUEST_JA_PATTERNS
      / RUST_REQUEST_PATTERNS / etc) that were accidentally deleted
      during the multi-range sed extraction (these are unrelated to
      Phase 2 and were collateral damage from a line-shift bug).

Metrics: turn.rs 34,881 -> 34,833 (-48); verifier_orchestration.rs
119 -> 181 (+62).
